Move remaining functions to output_formats.py
